+++ trigger-github-issues-from-crashlytics/functions/src/github_api.ts
@@ -1,7 +1,7 @@
-import { Octokit } from '@octokit/rest'
 import { RequestError } from '@octokit/request-error'
-import { CrashlyticsEvent } from 'firebase-functions/v2/alerts/crashlytics'
+import { Octokit } from '@octokit/rest'
 import { logger } from 'firebase-functions/v2'
+import { CrashlyticsEvent } from 'firebase-functions/v2/alerts/crashlytics'
 import { CrashlyticsAlert, CrashlyticsPayload } from './types'
 
 const octokit = new Octokit({
@@ -13,8 +13,9 @@ export async function createGitHubIssueIfEnabled<T extends CrashlyticsPayload>(
   event: CrashlyticsEvent<T>
 ) {
   const alertType = parseAlertType(event.alertType)
+  logger.info(`alertType: ${alertType}`, { structuredData: true })
   if (!process.env.ALERTS?.split(',').includes(alertType)) {
-    logger.info(
+    logger.warn(
       `Skip the creation of a GitHub issue because ${alertType} alert is not enabled`
     )
     return
@@ -42,9 +43,34 @@ export async function createGitHubIssueIfEnabled<T extends CrashlyticsPayload>(
   }
 }
 
+// Examople payload
+// {
+//   id: '2262421223909713056',
+//   appid: '1:162074331013:android:f5830f1b7bf6ffd8ade6ec',
+//   type: 'google.firebase.firebasealerts.alerts.v1.published',
+//   alerttype: 'crashlytics.regression',
+//   source: '//firebasealerts.googleapis.com/projects/162074331013',
+//   project: '162074331013',
+//   specversion: '1.0',
+//   time: '2025-05-19T00:44:38.808893Z',
+//   data: {
+//     '@type': 'type.googleapis.com/google.events.firebase.firebasealerts.v1.AlertData',
+//     createTime: '2025-05-19T00:44:38.808893Z',
+//     endTime: '2025-05-19T00:44:38.808893Z',
+//     payload: {
+//       issue: [Object],
+//       '@type': 'type.googleapis.com/google.events.firebase.firebasealerts.v1.CrashlyticsRegressionAlertPayload',
+//       type: 'fatal',
+//       resolveTime: '2025-05-18T07:00:00Z'
+//     }
+//   },
+//   traceparent: '00-407a4a49f8d10296ddf36a791a38df95-6f41c394d4759033-01',
+//   alertType: 'crashlytics.regression',
+//   appId: '1:162074331013:android:f5830f1b7bf6ffd8ade6ec'
+// }
 function makeBody<T extends CrashlyticsPayload>(event: CrashlyticsEvent<T>) {
   logger.info(event, { structuredData: true })
-  const issue = event.data.payload.issue
+  const { id, subtitle, appVersion } = event.data.payload.issue
   const alertType = parseAlertType(event.alertType)
   // TODO(tsuruoka): 本来はURLリンクを載せたいものの、`packageName`が取得できず`appId`のみなのでURLを生成できない問題
   // フォーマットは以下であることを確認したので、`packageName`が取得できるようになったらURLも表示できる
@@ -52,14 +78,14 @@ function makeBody<T extends CrashlyticsPayload>(event: CrashlyticsEvent<T>) {
   const appId = event.appId
 
   return `
-  ### ${issue.subtitle}
+  ### ${subtitle}
 
   | Info | Value |
   |--------|--------|
   | appId | ${appId} |
   | alertType | ${alertType} |
-  | id | ${issue.id} |
-  | appVersion | ${issue.appVersion} |
+  | id | ${id} |
+  | appVersion | ${appVersion} |
   `
 }