+++ trigger-github-issues-from-crashlytics/functions/src/index.ts
@@ -2,12 +2,14 @@
 // ref. https://firebase.google.com/docs/extensions/publishers/functions#crashlytics
 import {
   CrashlyticsOptions,
-  onNewFatalIssuePublished,
-  onNewNonfatalIssuePublished,
-  onNewAnrIssuePublished,
+  NewAnrIssuePayload,
   NewFatalIssuePayload,
   NewNonfatalIssuePayload,
-  NewAnrIssuePayload,
+  onNewAnrIssuePublished,
+  onNewFatalIssuePublished,
+  onNewNonfatalIssuePublished,
+  onRegressionAlertPublished,
+  RegressionAlertPayload,
 } from 'firebase-functions/v2/alerts/crashlytics'
 import { createGitHubIssueIfEnabled } from './github_api'
 
@@ -15,6 +17,9 @@ const options: CrashlyticsOptions = {
   region: process.env.LOCATION,
   secrets: ['GITHUB_ACCESS_TOKEN'],
 }
+
+// 関数名は`4-63 characters`が制限なので、Extensionsの場合は以下の形式となるので簡素な名前をつけるのがベター
+// ext-trigger-github-issues-from-crashlytics-[FUNCTION_NAME]
 
 // New fatal issue
 exports.createFatalIssue = onNewFatalIssuePublished(options, (event) =>
@@ -29,4 +34,9 @@ exports.createNonFatalIssue = onNewNonfatalIssuePublished(options, (event) =>
 // New ANR issue
 exports.createAnrIssue = onNewAnrIssuePublished(options, (event) =>
   createGitHubIssueIfEnabled<NewAnrIssuePayload>(event)
+)
+
+// Regression(Better use for report debugging)
+exports.regressionAlert = onRegressionAlertPublished(options, (event) =>
+  createGitHubIssueIfEnabled<RegressionAlertPayload>(event)
 )