+++ trigger-github-issues-from-crashlytics/CHANGELOG.md
@@ -1,3 +1,9 @@
+## Version 1.0.0
+
+- Add new trigger from `regression` alert
+- Update runtime to `nodejs22`
+- Update docs
+
 ## Version 0.0.4
 
 - Update docs